[REFACTOR] create cube_network.cpp
containing the remaining stuff from tools.cpp
